Change both slack and twitter hooks to return "True" when not enabled - as they have succeeded at not doing anything.
This also adds a "cli" hook specifically for printing to console, useful for debugging if both of the other hooks are disabled.

This might not be the best way to do this, but it should stop the queue from endlessly retrying to send slack/twitter messages when the services are disabled.